feat: groundwork for Steam Auth
This commit introduces basic functionality for redirecting a user to
Steam and back. Steam's response is verified and the user's SteamID is
extracted and can be encoded as a JWT.

The next steps would be to put this token into the user's cookies and
extend the database to store permissions for logged in users. We
probably want to do this with bitfields. Whether we store those
permissions in the JWT as well is unclear; it would mean that
invalidating those permissions is basically impossible. Setting a very
short expiration date would not replace the ability to invalidate
immediately, and just be annoying for users, so we probably just want to
hit the database once we know a user is authenticated, even if that is
on every request. We'll have to see.