Change map creation / update logic (#24)
This PR changes the way maps and changes to maps are handled.

# Submitting Maps

If a player has created a map and wants it to be included in the global map
pool, they can submit it to the Map Approval Team for review. The map is either
accepted, or rejected with justification + feedback. The mapper can then decide
to either follow that feedback and continue the process, or not.

# Changing Maps

Once a map is accepted, no more gameplay can be added (e.g. via new courses),
and no existing gameplay can be modified. What _is_ possible is that mappers can
request either individual courses, or their entire map, to be removed from the
global map pool. In the first case it would mean changing the map itself to
remove that course. In the second case, the map is simply removed from the pool.

If bugs or "map-breaking" skips are found after the map has been released, the
Map Approval Team will ask the mapper to fix those issues. They will either
comply, or find someone else to fix them and give explicit permission. If they
decide not to do that, their map will be removed from the global map pool.

If the mapper decides to do a visual overhaul of their map, or someone organizes
a contest for revamping existing maps, then the Map Approval Team can decide to
accept those changes, as long as they do not affect gameplay in any way. This
includes both the gameplay itself (e.g. distances between blocks), but also
visuals that could make the map more difficult to play (e.g. new invisible
walls, or bad visiblity with new lighting / textures).

Once a map is accepted into the global map pool, any changes to the map will
automatically de-global the map. If this happens accidentally, no harm is
done. If it happens intentionally, the map might be degloballed (if the Map
Approval Team decides so).