refactor: enforce depotdownloader in production
Previously, the environment variables for where to store Workshop
assets, and the DepotDownloader binary, were optional so you could more
easily run the API locally. The existance of those variables should be
enforced in production though, and at that point their type shouldn't be
`Option<T>` anymore, and we shouldn't need to assert that they exist
more than once.